[Refactor] 백엔드 리팩토링
올바르지 않은 경로 수정 및 에러 처리

#155
