+++ backend/src/route/cloud.ts
@@ -23,8 +23,13 @@ import {
 	getNewFolder,
 	getTrashFiles,
 	updateDir,
+	splitFolderAndFile,
+	FilteredFilesArg,
+	FilesArg,
+	getFilteredFiles,
+	getFiles,
 } from '../service/cloud';
-
+import { applyEscapeString } from '../util';
 const router = express.Router();
 
 router.get('/validate', isAuthenticated, async (req, res) => {
@@ -43,11 +48,14 @@ router.post('/upload', isAuthenticated, upload.array('uploadFiles'), async (req,
 	const files = req.files as Express.Multer.File[];
 	const { loginId } = req.user;
 	const { relativePath, rootDirectory } = req.body;
-	const relativePaths = JSON.parse(relativePath);
+	if (loginId === undefined || relativePath === undefined || rootDirectory === undefined) {
+		return res.status(400).send();
+	}
 
+	const relativePaths = JSON.parse(relativePath);
 	try {
-		const totalSzie = files.reduce((prev, file) => prev + file.size, 0);
-		if (!(await canIncreaseCurrentCapacity({ loginId: loginId, value: totalSzie }))) {
+		const totalSize = files.reduce((prev, file) => prev + file.size, 0);
+		if (!(await canIncreaseCurrentCapacity({ loginId: loginId, value: totalSize }))) {
 			return res.status(403).send();
 		}
 
@@ -71,15 +79,16 @@ router.post('/upload', isAuthenticated, upload.array('uploadFiles'), async (req,
 			})
 		);
 
-		res.status(200).send();
+		return res.status(200).send();
 	} catch (err) {
-		res.status(500).send();
+		return res.status(500).send();
 	}
 });
 
 router.get('/download', isAuthenticated, async (req, res) => {
 	const { loginId } = req.user;
 	const { current_dir, files, folders } = req.query;
+
 	if (current_dir === undefined || files === undefined || folders === undefined) {
 		return res.status(400).send();
 	}
@@ -96,7 +105,9 @@ router.get('/download', isAuthenticated, async (req, res) => {
 
 	createZipFile({ targetFolderPath, zipFolderPath });
 	res.download(zipFolderPath, `${loginId}.zip`, (err) => {
-		if (!err) {
+		if (err) {
+			res.status(500).send();
+		} else {
 			res.send();
 		}
 	});
@@ -106,8 +117,14 @@ router.get('/download', isAuthenticated, async (req, res) => {
 });
 
 router.put('/files', isAuthenticated, async (req, res) => {
-	const { targetIds = [], directories = [], action } = req.body;
 	const { loginId } = req.user;
+	const { targetIds = [], directories = [], action } = req.body;
+	if (loginId === undefined) {
+		return res.status(400).send();
+	}
+	if (targetIds === undefined || directories === undefined || action === undefined) {
+		return res.status(400).send();
+	}
 	try {
 		switch (action) {
 			case FileEditAction.trash:
@@ -131,10 +148,9 @@ router.put('/files', isAuthenticated, async (req, res) => {
 				});
 				break;
 		}
-
-		res.send();
+		return res.status(200).send();
 	} catch (err) {
-		res.status(500).send();
+		return res.status(500).send();
 	}
 });
 
@@ -183,4 +199,64 @@ router.post('/update', isAuthenticated, async (req, res) => {
 	return res.send();
 });
 
+router.get('/files', isAuthenticated, async (req, res) => {
+	const { path, isAscending, isDeleted, isStar } = req.query;
+	const { loginId } = req.user;
+	if (path === undefined || isAscending === undefined || isDeleted === undefined) {
+		return res.status(400).send();
+	}
+	if (path === '') {
+		return res.status(400).send();
+	}
+
+	try {
+		const filesArg: FilesArg = {
+			loginId: loginId,
+			regex: `(^${applyEscapeString(path as string)}$)|(^${
+				path === '/' ? '' : applyEscapeString(path as string)
+			}/(.*)?$)`,
+			isAscending: isAscending === 'true',
+			isDeleted: isDeleted === 'true',
+			isStar: isStar === 'true',
+		};
+
+		const tempFiles = await getFiles(filesArg);
+
+		const filteredFilesArg: FilteredFilesArg = {
+			path: path as string,
+			originFiles: tempFiles,
+		};
+
+		const files = getFilteredFiles(filteredFilesArg);
+		return res.status(200).json(files);
+	} catch (err) {
+		return res.status(500).send();
+	}
+});
+
+router.get('/starfiles', isAuthenticated, async (req, res) => {
+	const { path, isAscending } = req.query;
+	const { loginId } = req.user;
+	if (path === undefined || isAscending === undefined) {
+		return res.status(400).send();
+	}
+	if (path === '') {
+		return res.status(400).send();
+	}
+
+	const filesArg: FilesArg = {
+		loginId: loginId,
+		regex: `.+`,
+		isAscending: isAscending === 'true',
+		isDeleted: false,
+		isStar: true,
+	};
+	try {
+		const tempFiles = await getFiles(filesArg);
+		const data = splitFolderAndFile(tempFiles);
+		return res.status(200).json(data);
+	} catch (err) {
+		return res.status(500).send();
+	}
+});
 export default router;
